Update RomajiInput import paths and rename calcFirstAlphabetOfOkurigana function to lookupOkuriAlphabet
