+++ src/input-mode/AbstractKanaMode.ts
@@ -1,6 +1,6 @@
 import * as vscode from 'vscode';
 import * as wanakana from 'wanakana';
-import { RomajiInput } from '../RomajiInput';
+import { RomajiInput } from '../lib/romaji/RomajiInput';
 import { insertOrReplaceSelection, replaceRange, setInputMode } from '../extension';
 import { Candidate } from '../jisyo/candidate';
 import { AbstractInputMode } from './AbstractInputMode';
