+++ src/input-mode/henkan/KakuteiMode.ts
@@ -1,7 +1,7 @@
 import * as vscode from 'vscode';
 import { AbstractKanaMode } from '../AbstractKanaMode';
 import { AbstractHenkanMode } from './AbstractHenkanMode';
-import { RomajiInput } from '../../RomajiInput';
+import { RomajiInput } from '../../lib/romaji/RomajiInput';
 import { MidashigoMode } from './MidashigoMode';
 import { AbbrevMode } from './AbbrevMode';
 import { setInputMode } from '../../extension';
