+++ src/input-mode/henkan/MidashigoMode.ts
@@ -1,4 +1,4 @@
-import { RomajiInput } from "../../RomajiInput";
+import { RomajiInput } from "../../lib/romaji/RomajiInput";
 import { DeleteLeftResult, IEditor } from "../../editor/IEditor";
 import { insertOrReplaceSelection, setInputMode } from "../../extension";
 import { Entry } from "../../jisyo/entry";
@@ -10,6 +10,7 @@ import { AbstractMidashigoMode } from "./AbstractMidashigoMode";
 import { InlineHenkanMode } from "./InlineHenkanMode";
 import { KakuteiMode } from "./KakuteiMode";
 import { openRegistrationEditor } from './RegistrationEditor';
+import { lookupOkuriAlphabet } from "../../jisyo/okuri";
 
 export enum MidashigoType {
     gokan, // ▽あ
@@ -42,7 +43,7 @@ export class MidashigoMode extends AbstractMidashigoMode {
     }
 
     private createJisyoKey(midashigo: string, okuri: string): {key: string, keyForLookup: string} {
-        const okuriAlphabet = okuri.length > 0 ? (calcFirstAlphabetOfOkurigana(okuri) || "") : "";
+        const okuriAlphabet = okuri.length > 0 ? (lookupOkuriAlphabet(okuri) || "") : "";
         const key = midashigo + okuriAlphabet;
         const keyForLookup = this.romajiInput.convertKanaToHiragana(key);
         return {key, keyForLookup};
@@ -62,7 +63,7 @@ export class MidashigoMode extends AbstractMidashigoMode {
             return;
         }
 
-        const okuriAlphabet = okuri.length > 0 ? (calcFirstAlphabetOfOkurigana(okuri) || "") : "";
+        const okuriAlphabet = okuri.length > 0 ? (lookupOkuriAlphabet(okuri) || "") : "";
         context.setHenkanMode(new InlineHenkanMode(context, this.editor, this, midashigo, okuriAlphabet, jisyoEntry, optionalSuffix));
     }
 
@@ -217,155 +218,3 @@ export class MidashigoMode extends AbstractMidashigoMode {
     }
 }
 
-function calcFirstAlphabetOfOkurigana(okurigana: string): string | undefined {
-    return kanaToAlphabet.get(okurigana[0]);
-}
-
-const kanaToAlphabet = new Map<string, string>([
-    ["あ", "a"],
-    ["い", "i"],
-    ["う", "u"],
-    ["え", "e"],
-    ["お", "o"],
-    ["か", "k"],
-    ["が", "g"],
-    ["き", "k"],
-    ["ぎ", "g"],
-    ["く", "k"],
-    ["ぐ", "g"],
-    ["け", "k"],
-    ["げ", "g"],
-    ["こ", "k"],
-    ["ご", "g"],
-    ["さ", "s"],
-    ["ざ", "z"],
-    ["し", "s"],
-    ["じ", "z"],
-    ["す", "s"],
-    ["ず", "z"],
-    ["せ", "s"],
-    ["ぜ", "z"],
-    ["そ", "s"],
-    ["ぞ", "z"],
-    ["た", "t"],
-    ["だ", "d"],
-    ["ち", "t"],
-    ["ぢ", "d"],
-    ["つ", "t"],
-    ["づ", "d"],
-    ["て", "t"],
-    ["で", "d"],
-    ["と", "t"],
-    ["ど", "d"],
-    ["な", "n"],
-    ["に", "n"],
-    ["ぬ", "n"],
-    ["ね", "n"],
-    ["の", "n"],
-    ["は", "h"],
-    ["ば", "b"],
-    ["ぱ", "p"],
-    ["ひ", "h"],
-    ["び", "b"],
-    ["ぴ", "p"],
-    ["ふ", "h"],
-    ["ぶ", "b"],
-    ["ぷ", "p"],
-    ["へ", "h"],
-    ["べ", "b"],
-    ["ぺ", "p"],
-    ["ほ", "h"],
-    ["ぼ", "b"],
-    ["ぽ", "p"],
-    ["ま", "m"],
-    ["み", "m"],
-    ["む", "m"],
-    ["め", "m"],
-    ["も", "m"],
-    ["や", "y"],
-    ["ゆ", "y"],
-    ["よ", "y"],
-    ["ら", "r"],
-    ["り", "r"],
-    ["る", "r"],
-    ["れ", "r"],
-    ["ろ", "r"],
-    ["わ", "w"],
-    ["ゐ", "w"],
-    ["ゑ", "w"],
-    ["を", "w"],
-    ["ん", "n"],
-    ["ア", "a"],
-    ["イ", "i"],
-    ["ウ", "u"],
-    ["エ", "e"],
-    ["オ", "o"],
-    ["カ", "k"],
-    ["ガ", "g"],
-    ["キ", "k"],
-    ["ギ", "g"],
-    ["ク", "k"],
-    ["グ", "g"],
-    ["ケ", "k"],
-    ["ゲ", "g"],
-    ["コ", "k"],
-    ["ゴ", "g"],
-    ["サ", "s"],
-    ["ザ", "z"],
-    ["シ", "s"],
-    ["ジ", "z"],
-    ["ス", "s"],
-    ["ズ", "z"],
-    ["セ", "s"],
-    ["ゼ", "z"],
-    ["ソ", "s"],
-    ["ゾ", "z"],
-    ["タ", "t"],
-    ["ダ", "d"],
-    ["チ", "t"],
-    ["ヂ", "d"],
-    ["ツ", "t"],
-    ["ヅ", "d"],
-    ["テ", "t"],
-    ["デ", "d"],
-    ["ト", "t"],
-    ["ド", "d"],
-    ["ナ", "n"],
-    ["ニ", "n"],
-    ["ヌ", "n"],
-    ["ネ", "n"],
-    ["ノ", "n"],
-    ["ハ", "h"],
-    ["バ", "b"],
-    ["パ", "p"],
-    ["ヒ", "h"],
-    ["ビ", "b"],
-    ["ピ", "p"],
-    ["フ", "h"],
-    ["ブ", "b"],
-    ["プ", "p"],
-    ["ヘ", "h"],
-    ["ベ", "b"],
-    ["ペ", "p"],
-    ["ホ", "h"],
-    ["ボ", "b"],
-    ["ポ", "p"],
-    ["マ", "m"],
-    ["ミ", "m"],
-    ["ム", "m"],
-    ["メ", "m"],
-    ["モ", "m"],
-    ["ヤ", "y"],
-    ["ユ", "y"],
-    ["ヨ", "y"],
-    ["ラ", "r"],
-    ["リ", "r"],
-    ["ル", "r"],
-    ["レ", "r"],
-    ["ロ", "r"],
-    ["ワ", "w"],
-    ["ヰ", "w"],
-    ["ヱ", "w"],
-    ["ヲ", "w"],
-    ["ン", "n"],
-]);
